+++ lib/blocks/document.dart
@@ -12,12 +12,7 @@ class Document {
 
   /// Inserts a new [paragraph] into the document.
   void insert(Paragraph paragraph) {
-    final formattedPrs = _formatParagraph(paragraph);
-    if (formattedPrs.isEmpty) {
-      paragraphs.add(paragraph);
-      return;
-    }
-    paragraphs.addAll(formattedPrs);
+    paragraphs.add(paragraph);
   }
 
   /// Returns the last [paragraph] into the document and validate before to avoid exceptions.
@@ -27,13 +22,26 @@ class Document {
   }
 
   /// Returns the last [paragraph] into the document.
-  Paragraph? getLast() {
-    return paragraphs.lastOrNull;
+  Paragraph? getLast({Paragraph Function()? orElse}) {
+    return paragraphs.lastOrNull ?? orElse?.call();
   }
 
   /// Update a last [paragraph] into the document validating to make more safe the operation.
   void updateLastSafe(Paragraph paragraph) {
-    final int lastIndex = paragraphs.lastIndexOf(paragraph);
+    if(paragraphs.isEmpty) {
+      paragraphs.add(paragraph);
+      return;
+    }
+    paragraphs[paragraphs.length - 1] = paragraph;
+  }
+
+  /// Update a [paragraph] into the document validating to make more safe the operation.
+  void updateParagraph(Paragraph paragraph) {
+    int lastIndex = paragraphs.lastIndexOf(paragraph);
+    // make a second check to be sure that it does exist or not
+    if (lastIndex == -1 && paragraphs.isNotEmpty) {
+      lastIndex = paragraphs.indexWhere((pr) => pr.id == paragraph.id);
+    }
     if (paragraphs.isEmpty || lastIndex == -1) {
       paragraphs.add(paragraph);
       return;
@@ -51,36 +59,6 @@ class Document {
     paragraphs.clear();
   }
 
-  Iterable<Paragraph> _formatParagraph(Paragraph paragraph) {
-    final List<Paragraph> newParagraphs = [];
-    if (paragraph.lines.isNotEmpty) {
-      final Line line = paragraph.lines.first;
-      if (line.data == '\n' && paragraph.lines.length > 1) {
-        newParagraphs.add(
-          Paragraph(
-            lines: [Line(data: '\n')],
-            type: ParagraphType.lineBreak,
-          ),
-        );
-        paragraph.removeLine(0);
-        if (paragraph.lines.isNotEmpty) {
-          paragraph.setTypeSafe(paragraph.blockAttributes != null ? ParagraphType.block : ParagraphType.inline);
-          newParagraphs.add(paragraph.clone);
-        }
-      } else {
-        if (line.data == '\n' && paragraph.blockAttributes == null && paragraph.lines.length == 1) {
-          paragraph.setType(ParagraphType.lineBreak);
-        }
-        if (paragraph.blockAttributes != null) {
-          paragraph.setTypeSafe(ParagraphType.block);
-        } else if (paragraph.type != ParagraphType.lineBreak) {
-          paragraph.setTypeSafe(ParagraphType.inline);
-        }
-        newParagraphs.add(paragraph);
-      }
-    }
-    return newParagraphs;
-  }
 
   /// Ensures correct formatting of paragraphs in the document.
   @Deprecated('ensureCorrectFormat is no longer used and will be removed in future releases')
@@ -126,14 +104,14 @@ class Document {
   /// Returns a version of the string that can be readed more easily.
   String toPrettyString() {
     final StringBuffer buffer = StringBuffer('  Paragraph:\n');
-    return 'Document:\n'
-            '${paragraphs.map((Paragraph paragraph) {
+    final String rawParagraph = paragraphs.map((Paragraph paragraph) {
       for (final Line line in paragraph.lines) {
-        buffer.write('    $line\n');
+        buffer.writeln('    $line');
       }
-      final String attrStr =
-          paragraph.blockAttributes != null ? 'Paragraph Attributes: ${paragraph.blockAttributes}' : "";
-      final String typeStr = paragraph.type != null ? 'Type: ${paragraph.type?.name}' : '';
+      final String attrStr = paragraph.blockAttributes != null
+          ? 'Paragraph Attributes: ${paragraph.blockAttributes ?? <String, dynamic>{}}'
+          : "";
+      final String typeStr = 'Type: ${paragraph.type.name}';
       if (attrStr.isNotEmpty) {
         buffer.write('    $attrStr\n');
       }
@@ -145,8 +123,8 @@ class Document {
         ..clear()
         ..write('  Paragraph:\n');
       return str;
-    })}'
-        .replaceAll(RegExp(r',|\.|\(|\)'), '');
+    }).join();
+    return 'Document:\n$rawParagraph';
   }
 
   @override